+++ SFT/unsloth_compiled_cache/Conv1d.py
@@ -67,4 +67,4 @@ from typing import Any, List, Optional, Tuple, Union, Dict, Set, Callable
 from transformers.models.nemotron.modeling_nemotron import (Tensor)
 
 def forward(self, input: Tensor) -> Tensor:
-    return self._conv_forward(input, self.weight, self.bias).to(input.dtype).to(input.dtype).to(input.dtype).to(input.dtype).to(input.dtype)
+    return self._conv_forward(input, self.weight, self.bias).to(input.dtype).to(input.dtype).to(input.dtype).to(input.dtype).to(input.dtype).to(input.dtype)
