+++ SFT/unsloth_compiled_cache/LayerNorm.py
@@ -69,4 +69,4 @@ from transformers.models.nemotron.modeling_nemotron import (F, Tensor)
 def forward(self, input: Tensor) -> Tensor:
     return F.layer_norm(
         input, self.normalized_shape, self.weight, self.bias, self.eps
-    ).to(input.dtype).to(input.dtype).to(input.dtype).to(input.dtype).to(input.dtype)
+    ).to(input.dtype).to(input.dtype).to(input.dtype).to(input.dtype).to(input.dtype).to(input.dtype)
