+++ SFT/unsloth_compiled_cache/RMSNorm.py
@@ -70,4 +70,4 @@ def forward(self, x: torch.Tensor) -> torch.Tensor:
     """
     Runs the forward pass.
     """
-    return F.rms_norm(x, self.normalized_shape, self.weight, self.eps).to(input.dtype).to(input.dtype).to(input.dtype).to(input.dtype).to(input.dtype)
+    return F.rms_norm(x, self.normalized_shape, self.weight, self.eps).to(input.dtype).to(input.dtype).to(input.dtype).to(input.dtype).to(input.dtype).to(input.dtype)
